+++ R/EngineGraphQLGitLab.R
@@ -53,37 +53,34 @@ EngineGraphQLGitLab <- R6::R6Class(
       }
       group_cursor <- ""
       iterations_number <- round(orgs_count / 100)
-      orgs_list <- purrr::map(1:iterations_number, function(x) {
+      full_orgs_list <- list()
+      for (x in 1:iterations_number) {
         response <- self$gql_response(
           gql_query = self$gql_query$groups(),
           vars = list("groupCursor" = group_cursor)
         )
-        response <- set_graphql_error_class(response)
-        if (length(response$data$groups$edges) == 0) {
-          class(response) <- c(class(response), "graphql_error")
-        }
         if (!inherits(response, "graphql_error")) {
           if (output == "only_names") {
             orgs_list <- purrr::map(response$data$groups$edges, ~ .$node$fullPath)
           } else {
             orgs_list <- purrr::map(response$data$groups$edges, ~ .$node)
           }
-          group_cursor <<- response$data$groups$pageInfo$endCursor
-          return(orgs_list)
+          group_cursor <- response$data$groups$pageInfo$endCursor
+          full_orgs_list <- append(full_orgs_list, orgs_list)
         } else {
-          return(response)
+          full_orgs_list <- response
+          break
         }
-      }, .progress = TRUE) |>
-        purrr::list_flatten()
-      orgs_list <- handle_graphql_error(orgs_list, verbose)
-      if (!inherits(orgs_list, "graphql_error")) {
+      }
+      full_orgs_list <- private$handle_graphql_error(full_orgs_list, verbose)
+      if (!inherits(full_orgs_list, "graphql_error")) {
         if (output == "only_names") {
-          all_orgs <- unlist(orgs_list)
+          all_orgs <- unlist(full_orgs_list)
         } else if (output == "full_table") {
-          all_orgs <- orgs_list
+          all_orgs <- full_orgs_list
         }
       } else {
-        all_orgs <- orgs_list
+        all_orgs <- full_orgs_list
       }
       return(all_orgs)
     },
@@ -155,7 +152,7 @@ EngineGraphQLGitLab <- R6::R6Class(
         }
         full_repos_list <- append(full_repos_list, repos_list)
       }
-      full_repos_list <- handle_graphql_error(full_repos_list, verbose)
+      full_repos_list <- private$handle_graphql_error(full_repos_list, verbose)
       return(full_repos_list)
     },
 
@@ -194,7 +191,7 @@ EngineGraphQLGitLab <- R6::R6Class(
           full_repos_list <- append(full_repos_list, repos_list)
         }
       }
-      full_repos_list <- handle_graphql_error(full_repos_list, verbose)
+      full_repos_list <- private$handle_graphql_error(full_repos_list, verbose)
       return(full_repos_list)
     },
 
@@ -602,7 +599,6 @@ EngineGraphQLGitLab <- R6::R6Class(
           )
         )
       }
-      response <- set_graphql_error_class(response)
       return(response)
     },