+++ tests/testthat/test-01-get_repos-GitLab.R
@@ -131,52 +131,8 @@ test_that("get_repos_page handles properly a GraphQL query error", {
   mockery::stub(
     test_graphql_gitlab_priv$get_repos_page,
     "self$gql_response",
-    list(
-      "errors" = list(
-        list(
-          "message" = "Field 'count' doesn't exist on type 'ProjectConnection'",
-          "locations" = list(
-            list(
-              "line" = 6L,
-              "column" = 7L
-            )
-          ),
-          "path" = list(
-            "query GetReposByOrg",
-            "group",
-            "projects",
-            "count"
-          ),
-          "extensions" = list(
-            "code" = "undefinedField",
-            "typeName" = "ProjectConnection",
-            "fieldName" = "count"
-          )
-        ),
-        list(
-          "message" = "Field 'languages' doesn't exist on type 'Project'",
-          "locations" = list(
-            list(
-              "line" = 25L,
-              "column" = 11L
-            )
-          ),
-          "path" = list(
-            "query GetReposByOrg",
-            "group",
-            "projects",
-            "edges",
-            "node",
-            "languages"
-          ),
-          "extensions" = list(
-            "code" = "undefinedField",
-            "typeName" = "Project",
-            "fieldName" = "languages"
-          )
-        )
-      )
-    )
+    test_error_fixtures$graphql_error_no_count_languages |>
+      test_graphql_gitlab_priv$set_graphql_error_class()
   )
   repos_graphql_error <- test_graphql_gitlab_priv$get_repos_page()
   expect_s3_class(repos_graphql_error, "graphql_error")
@@ -184,7 +140,7 @@ test_that("get_repos_page handles properly a GraphQL query error", {
 })
 
 test_that("error handler works correctly", {
-  output <- handle_graphql_error(
+  output <- test_graphql_gitlab_priv$handle_graphql_error(
     responses_list = test_mocker$use("repos_graphql_error"),
     verbose = FALSE
   )
@@ -193,7 +149,7 @@ test_that("error handler works correctly", {
 
 test_that("error handler prints proper messages", {
   expect_snapshot(
-    output <- handle_graphql_error(
+    output <- test_graphql_gitlab_priv$handle_graphql_error(
       responses_list = test_mocker$use("repos_graphql_error"),
       verbose = TRUE
     )
@@ -517,6 +473,7 @@ test_that("parse_search_response works", {
   )
   gl_repos_raw_output <- gitlab_testhost_priv$parse_search_response(
     search_response = test_fixtures$gitlab_search_response,
+    org = "test_group",
     output = "raw"
   )
   expect_type(